Guard Async/Await with compiler version check
Also, rearrange some code into new files.
