+++ Sources/StructuredAPIClient/NetworkClient+AsyncAwait.swift
@@ -16,6 +16,8 @@ import Foundation
 import FoundationNetworking
 #endif
 
+#if compiler(>=5.5) && canImport(_Concurrency)
+
 @available(macOS 12.0, iOS 15.0, watchOS 8.0, tvOS 15.0, *)
 extension NetworkClient {
    public func load<Request: NetworkRequest>(_ req: Request) async throws -> Request.ResponseDataType {
@@ -27,3 +29,5 @@ extension NetworkClient {
         }
     }
 }
+
+#endif
