+++ Tests/StructuredAPIClientTests/NetworkClientWithAsyncAwaitTests.swift
@@ -18,6 +18,8 @@ import FoundationNetworking
 @testable import StructuredAPIClient
 import StructuredAPIClientTestSupport
 
+#if compiler(>=5.5) && canImport(_Concurrency)
+
 @available(macOS 12.0, iOS 15.0, watchOS 8.0, tvOS 15.0, *)
 final class NetworkClientWithAsyncAwaitTests: XCTestCase {
     
@@ -73,3 +75,5 @@ final class NetworkClientWithAsyncAwaitTests: XCTestCase {
         XCTAssertEqual(client.baseURL.absoluteString, "https://test.somewhere.com")
     }
 }
+
+#endif
